Make tests work without installing cookiecutter
When running tests without installing cookiecutter, some of them fail
because they try to invoke cookiecutter. Moreover, if cookiecutter is
installed on the system, the tests would use the installed version
instead of the one in the build tree.

To solve both issues, use `python -m cookiecutter.cli` instead of
`cookiecutter`.