Remove pull requests checklist reference
The link was broken since the file was removed from github gists. Its content
was recovered with Internet Archive's Wayback Machine and since it
significantly overlapped with what there's already in our
`CONTRIBUTING.md` it was just 'merged' into it. Credits to @audreyfeldroy as
original source of those ideas.

The part regarding `HISTORY.md` was removed since now the history is
updated using the output of Github's release drafter when we do a new
release.

Relates to #1528